Fix Fortran ordering with selection
When using column-major (Fortran) ordering, previously this was achieved
using a transpose of the original array. The intention was to use
row-major ordering internally for performance reasons. The problem is
that this effectively inverts the shape of the array. For example,
transposing a [5, 2] array yields a [2, 5] array. The selection would
then applied to the wrong dimensions. numpy is not strict about the end
of the slice, so in this case if requesting the first 5 of the outer
dimension this would yield only 2.

This change fixes the issue by reshaping the 1D array obtained from the
S3 data into the expected shape while at the same time applying the
ordering.